RYSK-26: Hoist modal to shared component.
