+++ packages/front-end/src/components/shared/RyskModal/types.ts
@@ -1,0 +1,5 @@
+import { PropsWithChildren } from "react";
+
+export interface RyskModalProps extends PropsWithChildren {
+  lightBoxClickFn: VoidFunction;
+}
